+++ peppol-lookup/src/main/java/no/difi/vefa/peppol/lookup/api/MetadataReader.java
@@ -2,6 +2,7 @@ package no.difi.vefa.peppol.lookup.api;
 
 import no.difi.vefa.peppol.common.model.DocumentIdentifier;
 import no.difi.vefa.peppol.common.model.ServiceMetadata;
+import no.difi.vefa.peppol.security.api.PeppolSecurityException;
 
 import java.util.List;
 
@@ -9,6 +10,6 @@ public interface MetadataReader {
 
     List<DocumentIdentifier> parseDocumentIdentifiers(FetcherResponse fetcherResponse) throws LookupException;
 
-    ServiceMetadata parseServiceMetadata(FetcherResponse fetcherResponse) throws LookupException, no.difi.vefa.peppol.security.api.SecurityException;
+    ServiceMetadata parseServiceMetadata(FetcherResponse fetcherResponse) throws LookupException, PeppolSecurityException;
 
 }