+++ peppol-security/src/main/java/no/difi/vefa/peppol/security/api/CertificateValidator.java
@@ -3,5 +3,5 @@ package no.difi.vefa.peppol.security.api;
 import java.security.cert.X509Certificate;
 
 public interface CertificateValidator {
-    void validate(X509Certificate certificate) throws SecurityException;
+    void validate(X509Certificate certificate) throws PeppolSecurityException;
 }