+++ peppol-security/src/main/java/no/difi/vefa/peppol/security/api/PeppolSecurityException.java
@@ -2,15 +2,15 @@ package no.difi.vefa.peppol.security.api;
 
 import no.difi.vefa.peppol.common.api.PeppolException;
 
-public class SecurityException extends PeppolException {
+public class PeppolSecurityException extends PeppolException {
 
     private static final long serialVersionUID = 6928682319726226728L;
 
-    public SecurityException(String message) {
+    public PeppolSecurityException(String message) {
         super(message);
     }
 
-    public SecurityException(String message, Throwable cause) {
+    public PeppolSecurityException(String message, Throwable cause) {
         super(message, cause);
     }
 }