+++ peppol-security/src/main/java/no/difi/vefa/peppol/security/xmldsig/XmldsigVerifier.java
@@ -1,6 +1,6 @@
 package no.difi.vefa.peppol.security.xmldsig;
 
-import no.difi.vefa.peppol.security.api.SecurityException;
+import no.difi.vefa.peppol.security.api.PeppolSecurityException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.w3c.dom.Document;
@@ -15,11 +15,11 @@ public class XmldsigVerifier {
 
     private static final Logger logger = LoggerFactory.getLogger(XmldsigVerifier.class);
 
-    public static X509Certificate verify(Document document) throws SecurityException {
+    public static X509Certificate verify(Document document) throws PeppolSecurityException {
         try {
             NodeList nl = document.getElementsByTagNameNS(XMLSignature.XMLNS, "Signature");
             if (nl.getLength() == 0)
-                throw new SecurityException("Cannot find Signature element");
+                throw new PeppolSecurityException("Cannot find Signature element");
 
             X509KeySelector keySelector = new X509KeySelector();
             DOMValidateContext valContext = new DOMValidateContext(keySelector, nl.item(0));
@@ -28,15 +28,15 @@ public class XmldsigVerifier {
             XMLSignature signature = xmlSignatureFactory.unmarshalXMLSignature(valContext);
 
             if (!signature.validate(valContext))
-                throw new SecurityException("Signature failed.");
+                throw new PeppolSecurityException("Signature failed.");
 
             logger.debug("Signature passed.");
             return keySelector.getCertificate();
-        } catch (SecurityException e) {
+        } catch (PeppolSecurityException e) {
             throw e;
         } catch (Exception e) {
             logger.warn(e.getMessage(), e);
-            throw new SecurityException("Unable to verify document signature.", e);
+            throw new PeppolSecurityException("Unable to verify document signature.", e);
         }
     }